Flyttar Adresse-klassa til modellen
