+++ apps/etterlatte-brev-api/src/main/kotlin/no/nav/etterlatte/brev/model/BrevModel.kt
@@ -10,30 +10,6 @@ import java.time.LocalDate
 import java.util.UUID
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-data class Adresse(
-    val adresseType: String,
-    val adresselinje1: String? = null,
-    val adresselinje2: String? = null,
-    val adresselinje3: String? = null,
-    val postnummer: String? = null,
-    val poststed: String? = null,
-    val landkode: String,
-    val land: String,
-) {
-    fun erGyldig(): Boolean {
-        return if (adresseType.isBlank() || landkode.isBlank() || land.isBlank()) {
-            false
-        } else if (adresseType == "NORSKPOSTADRESSE") {
-            !(postnummer.isNullOrBlank() || poststed.isNullOrBlank())
-        } else if (adresseType == "UTENLANDSKPOSTADRESSE") {
-            !adresselinje1.isNullOrBlank()
-        } else {
-            true
-        }
-    }
-}
-
-@JsonIgnoreProperties(ignoreUnknown = true)
 data class Mottaker(
     val navn: String,
     val foedselsnummer: Foedselsnummer? = null,
